+++ minhas-financas-back-end/MinhasFinancas.Application/Interfaces/ILancamentoAppService.cs
@@ -11,6 +11,8 @@ namespace MinhasFinancas.Application.Interfaces
         Task<RetornoGenerico> EfetivarLancamentoAsync(string usuarioId, Guid lancamentoId);
         Task<RetornoGenerico> BuscarFluxoCaixaSimplesAsync(string usuarioId, int ano, int mes);
         Task<RetornoGenerico> ExportarLancamentosExcelAsync(string usuarioId, FiltroListagemLancamentoDTO filtro);
+        Task<RetornoGenerico> BaixarModeloImportacaoLancamentosExcelAsync(string usuarioId);
+        Task<RetornoGenerico> ImportarLancamentosExcelAsync(string usuarioId, Stream arquivo);
         Task<RetornoGenerico> ExportarFluxoCaixaSimplesExcelAsync(string usuarioId, ExportarFluxoCaixaSimplesExcelDTO filtro);
         Task<RetornoGenerico> BuscarParcelamentoAsync(string usuarioId, Guid grupoParcelamentoId);
         Task<RetornoGenerico> EditarParcelamentoEmLoteAsync(string usuarioId, Guid grupoParcelamentoId, EditarParcelamentoEmLoteDTO dto);